fix: move GuardrailInspectionText to 'adapter' package
